Move IO-fixed dsl to org.http4s.dsl.io
