+++ blaze-server/src/test/scala/org/http4s/server/blaze/Http1ServerStageSpec.scala
@@ -9,7 +9,7 @@ import cats.effect._
 import org.http4s.blaze._
 import org.http4s.blaze.pipeline.{Command => Cmd}
 import org.http4s.blazecore.{ResponseParser, SeqTestHead}
-import org.http4s.dsl._
+import org.http4s.dsl.io._
 import org.http4s.headers.{Date, `Content-Length`, `Transfer-Encoding`}
 import org.http4s.{headers => H, _}
 import org.specs2.specification.core.Fragment
