+++ client/src/test/scala/org/http4s/client/middleware/FollowRedirectSpec.scala
@@ -7,7 +7,7 @@ import java.util.concurrent.atomic._
 import cats.effect._
 import cats.implicits._
 import fs2._
-import org.http4s.dsl._
+import org.http4s.dsl.io._
 import org.http4s.headers._
 import org.specs2.mutable.Tables
