+++ client/src/test/scala/org/http4s/client/middleware/RetrySpec.scala
@@ -5,7 +5,7 @@ package middleware
 import cats.effect.IO
 import cats.implicits._
 import fs2._
-import org.http4s.dsl._
+import org.http4s.dsl.io._
 import org.specs2.specification.Tables
 
 import scala.concurrent.duration._
