+++ docs/src/main/tut/client.md
@@ -29,7 +29,7 @@ Then we create the [service] again so tut picks it up:
 ```tut:book
 import cats.effect._
 import org.http4s._
-import org.http4s.dsl._
+import org.http4s.dsl.io._
 import org.http4s.server.blaze._
 
 val service = HttpService[IO] {
