+++ dsl/src/main/scala/org/http4s/dsl/io/package.scala
@@ -1,0 +1,5 @@
+package org.http4s.dsl
+
+import cats.effect.IO
+
+package object io extends Http4sDsl[IO]
