--- dsl/src/main/scala/org/http4s/dsl/package.scala
@@ -1,5 +1,0 @@
-package org.http4s
-
-import cats.effect.IO
-
-package object dsl extends Http4sDsl[IO]
