+++ dsl/src/test/scala/org/http4s/dsl/ResponseGeneratorSpec.scala
@@ -3,6 +3,7 @@ package dsl
 
 import cats._
 import cats.effect.IO
+import org.http4s.dsl.io._
 import org.http4s.headers.{Accept, `Content-Length`, `Content-Type`}
 
 class ResponseGeneratorSpec extends Http4sSpec {
