+++ examples/blaze/src/main/scala/com/example/http4s/blaze/ClientPostExample.scala
@@ -4,7 +4,7 @@ import cats.effect._
 import org.http4s._
 import org.http4s.client._
 import org.http4s.client.blaze.{defaultClient => client}
-import org.http4s.dsl._
+import org.http4s.dsl.io._
 
 object ClientPostExample extends App {
   val req = POST(uri("https://duckduckgo.com/"), UrlForm("q" -> "http4s"))
