+++ examples/src/main/scala/com/example/http4s/ExampleService.scala
@@ -9,7 +9,7 @@ import org.http4s.MediaType._
 import org.http4s._
 import org.http4s.server._
 import org.http4s.circe._
-import org.http4s.dsl._
+import org.http4s.dsl.io._
 import org.http4s.headers._
 import org.http4s.server.middleware.authentication.BasicAuth
 import org.http4s.twirl._
