+++ server/src/test/scala/org/http4s/server/middleware/ChunkAggregatorSpec.scala
@@ -7,7 +7,7 @@ import cats.instances.vector._
 import cats.syntax.foldable._
 import fs2._
 import org.http4s._
-import org.http4s.dsl._
+import org.http4s.dsl.io._
 import org.http4s.headers._
 import org.http4s.util.chunk._
 import org.scalacheck._
