+++ server/src/test/scala/org/http4s/server/middleware/DefaultHeadSpec.scala
@@ -4,7 +4,7 @@ package middleware
 
 import cats.effect._
 import fs2.Stream._
-import org.http4s.dsl._
+import org.http4s.dsl.io._
 
 class DefaultHeadSpec extends Http4sSpec {
   val service = DefaultHead[IO](HttpService[IO] {
