+++ server/src/test/scala/org/http4s/server/middleware/HSTSSpec.scala
@@ -3,7 +3,7 @@ package server
 package middleware
 
 import cats.effect._
-import org.http4s.dsl._
+import org.http4s.dsl.io._
 import org.http4s.headers._
 
 import scala.concurrent.duration._
