+++ server/src/test/scala/org/http4s/server/middleware/authentication/AuthenticationSpec.scala
@@ -12,7 +12,7 @@ import scala.concurrent.duration._
 
 import fs2._
 import org.http4s.Status._
-import org.http4s.dsl._
+import org.http4s.dsl.io._
 import org.http4s.headers._
 import org.http4s.parser.HttpHeaderParser
 import org.http4s.util.{ CaseInsensitiveString, NonEmptyList }
